Build KvikIO as a shared library (#527)
Moving libcurl dependent functions to `remote_handle.cpp` (let's move the rest in a follow up PR) and setup building `libkvikio.so`.

### Background 

Now that KvikIO has evolved into a standalone IO library that works without cuFile and CUDA, I think it makes sense to move to a shared library. Originally, KvikIO was a very thin wrapper around cuFile, which is why we decided to keep it header-only (not because of templates). 

**Pros**:
  - Enables us to statically compile libcurl into `libkvikio.so`. 
    * Required by our wheels. If we don’t do this, we will have to find another solution to https://github.com/rapidsai/kvikio/issues/512 
  - Avoid having to compile libcurl in downstream projects. Currently, cudf must compile libcurl in every build. 
  - Reduce compile time in CI, both for KvikIO and downstream projects. 
  - Ease development by not having to rebuild downstream projects like cudf every time KvikIO is modified. 

**Cons**: 
 - Projects cannot vendoring KvikIO as a header-only project. As far as we know, nobody does this.

Authors:
  - Mads R. B. Kristensen (https://github.com/madsbk)
  - Paul Taylor (https://github.com/trxcllnt)

Approvers:
  - Robert Maynard (https://github.com/robertmaynard)
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/kvikio/pull/527
