+++ cpp/include/kvikio/shim/utils.hpp
@@ -24,11 +24,10 @@
 namespace kvikio {
 
 // Macros used for defining symbol visibility.
-// Since KvikIO is header-only, we rely on the linker to disambiguate inline functions
-// and static methods that have (or return) static references. To do this, the relevant
-// function/method must have `__attribute__((visibility("default")))`. If not, then if
-// KvikIO is used in two different DSOs, the function will appear twice, and there will
-// be two static objects.
+// Since KvikIO declares global default values in headers, we rely on the linker to disambiguate
+// inline and static methods that have (or return) static references. To do this, the relevant
+// function/method must have `__attribute__((visibility("default")))`. If not, then if KvikIO is
+// used in two different DSOs, the function will appear twice, and there will be two static objects.
 // See <https://gcc.gnu.org/wiki/Visibility> and <https://github.com/rapidsai/kvikio/issues/442>.
 #if (defined(__GNUC__) || defined(__clang__)) && !defined(__MINGW32__) && !defined(__MINGW64__)
 #define KVIKIO_EXPORT __attribute__((visibility("default")))
