+++ dependencies.yaml
@@ -64,6 +64,7 @@ files:
     includes:
       - depends_on_cupy
       - depends_on_nvcomp
+      - depends_on_libkvikio
       - run
   py_rapids_build_libkvikio:
     output: pyproject
@@ -282,6 +283,31 @@ dependencies:
           - matrix:
             packages:
               - nvidia-nvcomp==4.1.0.6
+  depends_on_libkvikio:
+    common:
+      - output_types: conda
+        packages:
+          - &libkvikio_unsuffixed libkvikio==24.12.*,>=0.0.0a0
+      - output_types: requirements
+        packages:
+          # pip recognizes the index as a global option for the requirements.txt file
+          # This index is needed for libkvikio-cu{11,12}.
+          - --extra-index-url=https://pypi.nvidia.com
+          - --extra-index-url=https://pypi.anaconda.org/rapidsai-wheels-nightly/simple
+    specific:
+      - output_types: [requirements, pyproject]
+        matrices:
+          - matrix:
+              cuda: "12.*"
+              cuda_suffixed: "true"
+            packages:
+              - libkvikio-cu12==24.12.*,>=0.0.0a0
+          - matrix:
+              cuda: "11.*"
+              cuda_suffixed: "true"
+            packages:
+              - libkvikio-cu11==24.12.*,>=0.0.0a0
+          - {matrix: null, packages: [*libkvikio_unsuffixed]}
   docs:
     common:
       - output_types: [conda, requirements]
