+++ python/kvikio/pyproject.toml
@@ -20,6 +20,7 @@ license = { text = "Apache 2.0" }
 requires-python = ">=3.10"
 dependencies = [
     "cupy-cuda11x>=12.0.0",
+    "libkvikio==24.12.*,>=0.0.0a0",
     "numcodecs !=0.12.0",
     "numpy>=1.23,<3.0a0",
     "nvidia-nvcomp==4.1.0.6",
